More refactorizations related to game_ctrl and event modules: - Use STL containers, removing limitations for cups and races per event - Use size_t instead of int - Removed several unused variables - Removed garbage whitespaces - Fixed compilation with GCC (#include <algorithm> missing)
git-svn-id: https://svn.code.sf.net/p/extremetuxracer/code/trunk@311 0420edf4-82e4-42fc-9478-35b55e6d67a3

+++ gui.cpp
@@ -394,7 +394,7 @@ void DrawLevel (int x, int y, int level, double fact) {
 	glEnd();
 }
 
-void DrawBonus (int x, int y, int max, int num) {
+void DrawBonus (int x, int y, int max, size_t num) {
     TVector2 bl, tr;
 	double bott, top;
 	int i;
@@ -427,7 +427,7 @@ void DrawBonus (int x, int y, int max, int num) {
 	}
 }
 
-void DrawBonusExt (int y, int numraces, int num) {
+void DrawBonusExt (int y, size_t numraces, size_t num) {
     TVector2 bl, tr;
 	double bott, top;
 	int i;
